Moved three commented modules out of Common.hs; fixed one to actually build.
