+++ Language/Atom/Common.hs
@@ -153,70 +153,6 @@ readChannel (Channel a hasData) = do
   hasData <== false
   return a
 
--- | Fades one signal to another.
-module Language.Atom.Common.Fader
-  ( Fader
-  , FaderInit (..)
-  , fader
-  , fadeToA
-  , fadeToB
-  , fadeToCenter
-  ) where
-
-import Language.Atom
-
--- | Fader object.
-data Fader = Fader (V Int)
-
--- | Fader initalization.
-data FaderInit = OnA | OnB | OnCenter
-
-toA = 0
-toB = 1
-toCenter = 2
-
--- | Fader construction.  Name, fade rate, fader init, and signal A and B.
-fader :: Name -> Double -> FaderInit -> E Double -> E Double -> Atom (Fader, E Double)
-fader name rate init a b = scope name $ do
-  --assert "positiveRate" $ rate >= 0
-
-  target <- int (case init of {OnA -> toA; OnB -> toB; OnCenter -> toCenter})
-  perA <- double (case init of {OnA -> 1;   OnB -> 0;   OnCenter -> 0.5})
-
-  rule "toA" $ do
-    when $ value target ==. intC toA
-    when $ value perA <. 1
-    perA <== mux (1 - value perA <. doubleC rate) 1 (value perA + doubleC rate)
-
-  rule "toB" $ do
-    when $ value target ==. intC toB
-    when $ value perA >. 0
-    perA <== mux (value perA <. doubleC rate) 0 (value perA - doubleC rate)
-
-  rule "toCenterFrom0" $ do
-    when $ value target ==. intC toCenter
-    when $ value perA <. 0.5
-    perA <== mux (0.5 - value perA <. doubleC rate) 0.5 (value perA + doubleC rate)
-
-  rule "toCenterFrom1" $ do
-    when $ value target ==. intC toCenter
-    when $ value perA >. 0.5
-    perA <== mux (value perA - 0.5 <. doubleC rate) 0.5 (value perA - doubleC rate)
-
-  return (Fader target, (a * value perA + b * (1 - value perA)) / 2)
-
--- | Fade to signal A.
-fadeToA :: Fader -> Action ()
-fadeToA (Fader target) = target <== intC toA
-
--- | Fade to signal B.
-fadeToB :: Fader -> Action ()
-fadeToB (Fader target) = target <== intC toB
-
--- | Fade to center, ie average of signal A and B.
-fadeToCenter :: Fader -> Action ()
-fadeToCenter (Fader target) = target <== intC toCenter
-
 module Language.Atom.Common.Process
   ( Process (..)
   , process
@@ -233,93 +169,4 @@ data Process
 
 process :: Name -> Process -> Atom ()
 
--- | Time integrated threshold functions typically used in condition monitoring.
-module Language.Atom.Common.Threshold
-  ( boolThreshold
-  , floatingThreshold
-  ) where
-
-import Language.Atom
-
-
--- | Boolean thresholding over time.  Output is set when internal counter hits limit, and cleared when counter is 0.
-boolThreshold :: Name -> Int -> Bool -> E Bool -> Atom (E Bool)
-boolThreshold name num init input = scope name $ do
-  --assert "positiveNumber" $ num >= 0
-
-  state <- bool init
-  count <- int  (if init then num else 0)
-
-  rule "update" $ do
-    when $ value count >. 0 &&. value count <. num
-    count <== value count + mux input 1 (-1)
-
-  rule "low" $ do
-    when $ value count ==. 0
-    state <== false
-
-  rule "high" $ do
-    when $ value count ==. intC num
-    state <== true
-
-  return $ value state
-
--- | Integrating threshold.  Output is set with integral reaches limit, and cleared when integral reaches 0.
-doubleThreshold :: Name -> Double -> E Double -> Atom (E Bool)
-doubleThreshold name lim input = scope name $ do
-  --assert "positiveLimit" $ lim >= 0
-
-  state <- bool False
-  sum <- double 0
-
-  (high,low) <- priority
-
-  rule "update"
-    sum <== value sum + input
-    low
-
-  rule "clear" $ do
-    when $ value sum <=. 0
-    state <== false
-    sum <== 0
-    high
-
-  rule "set" $ do
-    when $ value sum >=. doubleC lim
-    state <== true
-    sum <== doubleC lim
-    high
-
-  return  $ value state
-
--- | Capturing data that can either be valid or invalid.
-module Language.Atom.Common.ValidData
-  ( ValidData
-  , validData
-  , getValidData
-  , whenValid
-  , whenInvalid
-  ) where
-
-import Language.Atom
-
--- | 'ValidData' captures the data and its validity condition.
---   'ValidData' is abstract to prevent rules from using invalid data.
-data ValidData a = ValidData a (E Bool)
-
--- | Create 'ValidData' given the data and validity condition.
-validData :: a -> E Bool -> ValidData a
-validData = ValidData
-
--- | Get a valid data.  Action is disabled if data is invalid.
-getValidData :: ValidData a -> Action a
-getValidData (ValidData a v) = cond v >> return a
-
--- | Action enabled if 'ValidData' is valid.
-whenValid :: ValidData a -> Action ()
-whenValid (ValidData _ v) = cond v
-
--- | Action enabled if 'ValidData' is not valid.
-whenInvalid :: ValidData a -> Action ()
-whenInvalid (ValidData _ v) = cond $ not_ v
 -}
